+++ src/Integration/ModulesIntegration.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Sentry\Integration;
 
+use Composer\InstalledVersions;
 use Jean85\PrettyVersions;
 use PackageVersions\Versions;
 use Sentry\Event;
@@ -63,11 +64,33 @@ final class ModulesIntegration implements IntegrationInterface
     private function processEvent(Event $event): void
     {
         if (empty(self::$loadedModules)) {
-            foreach (Versions::VERSIONS as $package => $rawVersion) {
-                self::$loadedModules[$package] = PrettyVersions::getVersion($package)->getPrettyVersion();
+            foreach (self::getInstalledPackages() as $package) {
+                try {
+                    self::$loadedModules[$package] = PrettyVersions::getVersion($package)->getPrettyVersion();
+                } catch (\Throwable $exception) {
+                    continue;
+                }
             }
         }
 
         $event->setModules(self::$loadedModules);
+    }
+
+    /**
+     * @return string[]
+     */
+    private static function getInstalledPackages(): array
+    {
+        if (class_exists(InstalledVersions::class)) {
+            return InstalledVersions::getInstalledPackages();
+        }
+
+        if (class_exists(Versions::class)) {
+            // BC layer for Composer 1, using a transient dependency
+            return array_keys(Versions::VERSIONS);
+        }
+
+        // this should not happen
+        return ['sentry/sentry'];
     }
 }