+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Allow jean85/pretty-package-versions 2.0 (#1171, backport of #1170 from 3.x)
+
 ### 2.5.1 (2020-12-11)
 
 - Fix regression introduced in #1129 (#1143)